006: AsyncHooks public API proposal
After much investigation and communication this is the API for
AsyncHooks that has evolved. Meant to be minimal, not impose any
performance penalty to core when not being used, and minimal impact when
it is used, this should serve public needs that have been expressed over
the last two years.

PR-URL: https://github.com/nodejs/node-eps/pull/18
Reviewed-By: Fedor Indutny <fedor@indutny.com>
Reviewed-By: Jeremiah Senkpiel <fishrock123@rocketmail.com>